Hydra/requirements.txt
absl-py==0.9.0
alabaster==0.7.12
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
decorator==4.4.2
Django==3.0.8
Flask==1.1.1
Flask-SQLAlchemy==2.4.4
Flask-Uploads==0.2.1
Flask-WTF==0.14.3
glob3==0.0.1
gunicorn==20.0.4
idna==2.10
image==1.5.32
imageio==2.9.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.2
networkx==2.4
numpy==1.17.3
opencv-python-headless==4.3.0.36
passlib==1.7.2
picterra==1.0.0
Pillow==7.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
requests==2.24.0
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.5.1
selenium==3.141.0
six==1.15.0
sklearn==0.0
SQLAlchemy==1.3.18
sqlparse==0.3.1
threadpoolctl==2.1.0
tifffile==2020.7.4
urllib3==1.25.9
Werkzeug==0.16.0
WTForms==2.3.1
